+++ prometheus/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.9.15
+
+* Setup for tls-client-validation
+
+---
+
 0.9.14
 
 * Turning off flow-control in syslog-ng, setting 120s time_reopen, and reducing log-fifo parameter